Extract runUserStylingMigration to own its failure reporting
Move the run-classify-report policy out of the atom-declaration module and into
migrateUserStyling.ts as an exported runUserStylingMigration that owns its own
catch/classify/markFailed. storageAtoms.ts now calls it in one line, dropping
three imports and the inline policy comment from what is otherwise a pure atom
declaration module.

Hoist the synthetic persistence-status key to a named MIGRATION_STATUS_KEY
const beside the migration, documenting that it is intentionally distinct from
any storage key so it is never auto-cleared. Fold the failure-path tests back
into migrateUserStyling.test.ts beside the rest of the migration tests; they
now drive the real function directly via a rejecting localForage spy instead of
mocking the module and re-importing storageAtoms.
