+++ packages/graph-explorer/src/core/StateProvider/storageAtoms.ts
@@ -1,5 +1,3 @@
-import { createErrorDetails } from "@/utils/createErrorDetails";
-
 import type {
   ConfigurationId,
   RawConfiguration,
@@ -14,32 +12,14 @@ import type {
 
 import { createActiveConfigurationAtom } from "./activeConnectionStorage";
 import { atomWithLocalForage } from "./atomWithLocalForage";
-import { migrateUserStylingIfNeeded } from "./migrateUserStyling";
-import { persistenceStatusStore } from "./persistence";
-import { classifyStorageError } from "./persistence/classifyStorageError";
+import { runUserStylingMigration } from "./migrateUserStyling";
 import { defaultUserLayout } from "./userLayoutDefaults";
 
 // Convert any legacy single-key user styling into the type-keyed map atoms
-// below before they preload. Must run before the `Promise.all`. Errors are
-// caught so a migration failure does not crash the module and leave all atoms
-// undefined. The legacy `"user-styling"` key is never deleted, so the data is
-// preserved on disk and the migration retries on the next reload. A migration
-// failure is a storage failure, so we report it through the same
-// persistence-status store as every other write failure: the "Changes not
-// saved" indicator lights up and its dialog shows this error alongside the
-// rest. The synthetic key never receives a successful write, so the failure
-// stays outstanding until the next reload clears it (or the migration
-// succeeds), which matches reality.
-try {
-  await migrateUserStylingIfNeeded();
-} catch (err) {
-  persistenceStatusStore.markFailed(
-    "user-styling-migration",
-    classifyStorageError(err),
-    1,
-    createErrorDetails(err),
-  );
-}
+// below before they preload. Must run before the `Promise.all` so the atoms
+// preload the migrated data. The migration owns its own failure reporting (a
+// failure surfaces through the persistence-status store), so it never throws.
+await runUserStylingMigration();
 
 /**
  DEV NOTE
